refactor: extract icons into subfolder
