+++ frontend/src/components/index.ts
@@ -1,8 +1,4 @@
 export { default as Avatar } from './Avatar';
 export { default as PinkButton } from './PinkButton';
-export { default as ExclamationCircleOutline } from './ExclamationCircleOutline';
-export { default as Exit } from './Exit';
-export { default as ProfileCircle } from './ProfileCircle';
 export { default as LegalText } from './LegalText';
 export { default as ErrorMessage } from './ErrorMessage';
-export { default as ArrowLeft } from './ArrowLeft';
