+++ frontend/src/icons/ExclamationCircleOutline.tsx
