+++ frontend/src/icons/Exit.tsx
