+++ frontend/src/icons/index.ts
@@ -1,0 +1,4 @@
+export { default as ProfileCircle } from './ProfileCircle';
+export { default as ExclamationCircleOutline } from './ExclamationCircleOutline';
+export { default as Exit } from './Exit';
+export { default as ArrowLeft } from './ArrowLeft';
